requirements.txt
dj-database-url==0.5.0
Django==2.2.5
django-cors-headers==3.1.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
gunicorn==19.9.0
psycopg2==2.8.3
PyJWT==1.7.1
pytz==2019.2
sqlparse==0.3.0
whitenoise==4.1.4
